+++ src/ExpenseTracker.Models/Dto/DtoExpenseFilter.cs
@@ -5,7 +5,7 @@ namespace ExpenseTracker.Models.Dto
     {
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int? Amount { get; set; }
+        public decimal? Amount { get; set; }
         public int CategoryId { get; set; }
         public int ExpenseTypeId { get; set; }
         public string Comment { get; set; }
